ts: publish the assembler contracts from ports and typecheck the emitted declarations
