+++ sdk-libs/ts/package.json
@@ -61,8 +61,9 @@
     "lint": "oxlint --deny-warnings src test api/test transaction/test wallet/test scripts",
     "test": "vitest run --exclude \"test/e2e/**\"",
     "test:e2e": "vitest run --no-file-parallelism test/e2e",
-    "check": "npm run format:check && npm run lint && npm run changelog:check && npm run typecheck && npm run test && npm run build",
-    "loadtest": "npm run build && node bench/loadtest.ts"
+    "check": "npm run format:check && npm run lint && npm run changelog:check && npm run typecheck && npm run test && npm run build && npm run check:dist",
+    "loadtest": "npm run build && node bench/loadtest.ts",
+    "check:dist": "tsc -p tsconfig.dist.json"
   },
   "dependencies": {
     "@lightprotocol/hasher.rs": "^0.2.1",
